requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==2.2.3
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==1.5.3
